+++ src/main/game/entities/units/Melee.java
@@ -4,28 +4,8 @@ import game.entities.*;
 import game.gameboard.Location;
 import game.visitors.iTileActionVisitor;
 
-public class Melee extends Unit implements iEntity, iAttacker, iDefender, iMoveable {
+public class Melee extends Unit implements iAttacker, iDefender, iMoveable {
 
     public Melee(UnitStats stats, Location location){ super(stats, location); }
 
-    /* iEntity */
-    public double getCurrentHealth(){ return this.health; }
-    public Percentage getHealthPercentage(){ return this.healthPercent; }
-    public void takeDamage(double damage){
-        this.health -= damage;
-        this.healthPercent.updateHealthPercentage((double)this.health);
-    }
-    public void heal(double healing){
-        this.health += healing;
-        this.healthPercent.updateHealthPercentage((double)this.health);
-    }
-    public void accept(iTileActionVisitor v){}
-
-
-    /* iAttacker */
-    public double getDamage(){ return (double)stats.getOffPow(); }
-    public int getRange(){ return 1; }
-
-    /* iMoveable */
-    public int getMoveDistance(){ return stats.getSpeed(); }
 }
